+++ lib/IE/component/IComponent/IComponentLayout/IComponentObject/IComponentObject.cpp
@@ -31,7 +31,7 @@ namespace ie {
 		return get_object().get_normal_size();
 	}
 	
-	void IComponentObject::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void IComponentObject::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		IComponent::draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		get_object().draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
 	}
