+++ lib/IE/component/IComponent/IComponentLayout/IComponentObjectsArray/IComponentObjectsArray.cpp
@@ -32,7 +32,7 @@ namespace ie {
 		return normal_size;
 	}
 	
-	void IComponentObjectsArray::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void IComponentObjectsArray::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		IComponent::draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		for(size_t i = 0; i < get_array_size(); ++i) {
 			get_object_at(i).draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
