+++ lib/IE/component/IComponent/IScalable/BasicInteractiveData/BaseButton/BaseButton.cpp
@@ -35,7 +35,7 @@ namespace ie {
 		return layout;
 	}
 	
-	void BaseButton::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BaseButton::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		background->draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 	}
 }
