+++ lib/IE/component/IComponent/IScalable/BasicInteractiveData/BaseButton/BaseButton.hpp
@@ -19,7 +19,7 @@ namespace ie {
 		
 		bool update_interactions(sf::Vector2f mouse_position) override;
 		
-		void draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) override;
+		void draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) override;
 	
 	protected:
 		LayoutData& layout_get_data() override;
