+++ lib/IE/component/IComponent/IScalable/BasicInteractiveData/BaseSlider/BaseSlider.cpp
@@ -133,7 +133,7 @@ namespace ie {
 		return true;
 	}
 	
-	void BaseSlider::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BaseSlider::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		background->draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		slider->draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
 	}
