+++ lib/IE/component/IComponent/IScalable/Box/BoxBorder/BoxBorder.cpp
@@ -138,7 +138,7 @@ namespace ie {
 		return new BoxBorder{*this};
 	}
 	
-	void BoxBorder::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BoxBorder::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		IComponent::draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		for(auto& line: objects) {
 			for(auto& object: line) {
