+++ lib/IE/component/IComponent/IScalable/Box/BoxBorderHorizontal/BoxBorderHorizontal.cpp
@@ -111,7 +111,7 @@ namespace ie {
 		return new BoxBorderHorizontal{*this};
 	}
 	
-	void BoxBorderHorizontal::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BoxBorderHorizontal::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		IComponent::draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		for(auto& object: objects) {
 			object->draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
