+++ lib/IE/component/IComponent/IScalable/Box/BoxBorderVertical/BoxBorderVertical.cpp
@@ -140,7 +140,7 @@ namespace ie {
 		return true;
 	}
 	
-	void BoxBorderVertical::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BoxBorderVertical::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		IComponent::draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		for(auto& object: objects) {
 			object->draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
