+++ lib/IE/component/IComponent/IScalable/Box/BoxConstRatio/BoxConstRatio.cpp
@@ -187,7 +187,7 @@ namespace ie {
 		return new BoxConstRatio{*this};
 	}
 	
-	void BoxConstRatio::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BoxConstRatio::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		IComponent::draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		const_object->draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
 		second_object->draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
