+++ lib/IE/component/IComponent/IScalable/Box/BoxPanel/BoxPanel.cpp
@@ -51,7 +51,7 @@ namespace ie {
 		return new BoxPanel{*this};
 	}
 	
-	void BoxPanel::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BoxPanel::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		object->draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		panel->draw_debug(render_target, 0, indent_addition, hue, hue_offset);
 	}
