+++ lib/IE/component/IComponent/IScalable/Box/BoxRenderTexture/BoxRenderTexture.cpp
@@ -95,7 +95,7 @@ namespace ie {
 		return new BoxRenderTexture{*this};
 	}
 	
-	void BoxRenderTexture::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BoxRenderTexture::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		object->draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 	}
